engine/engines/requirements_org.txt
Flask==2.0.3
flask-restful==0.3.9
Werkzeug==2.0.3
flask-restx==0.5.1
flask-cors==3.0.10
itsdangerous==2.0.1
pytest==7.0.1
sqlalchemy~=1.4.34
python-decouple==3.6
google-cloud-secret-manager==1.0.0
shapely==1.8.0
geoalchemy2==0.11.1
pandas==1.4.2
geopandas==0.11.0
psycopg2==2.9.1
gunicorn==20.1.0
tqdm==4.63.1
country_converter==0.8.0
pymongo[srv]==4.0.2
slack-logger==0.3.1
protobuf==3.20
scikit-image==0.19.3
rasterio==1.3.6
forex-python==1.8
beautifulsoup4==4.11.1
flask-mail==0.9.1
requests~=2.28.1
numpy~=1.23.0
pytz~=2022.1
pyproj~=3.3.1
selenium~=4.4.3
matplotlib~=3.5.2
lxml~=4.9.1
webdriver_manager~=3.8.3
seaborn~=0.12.1
slack-sdk==3.19.5
google-cloud-storage==2.7.0
pyreadr==0.4.7
pre-commit
kpler.sdk==1.0.29
Unidecode==1.3.6
pygeos==0.14
